app/backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
drf-spectacular==0.27.1

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2

# Database
mysqlclient==2.2.1
# psycopg2-binary==2.9.9  # Uncomment for PostgreSQL

# Geospatial support
# django-geojson==4.1.0
# Uncomment above when ready to use GeoDjango

# Image handling
Pillow==10.2.0

# WebSocket support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Development tools
python-decouple==3.8
